Crate foundationdb_simulation

FoundationDB-workload module

This module provides all necessary bindings for a FoundationDB’s ExternalWorkload under a Rust trait, as well as a way to register a Workload in the simulation.

Macros

details

Macro that can be used to create Details more easily.

Structs

Metric

A single metric entry

Promise

A wrapper around a FoundationDB promise

WorkloadContext

A wrapper around a FoundationDB promise

Enums

Severity

Indicates the severity of a FoundationDB log entry

Traits

RustWorkload

RustWorkload trait provides a one to one equivalent to the C++ abstract class FDBWorkload

Functions

CPPWorkloadFactory⚠

Redirect FoundationDB call to symbol “workloadFactory” (in Rust) to the C++ symbol “CPPWorkloadFactory”, it instantiate and returns a pointer to the C++ class RustWorkloadFactory.

fdb_spawn

Spawn an async block and resolve all contained FoundationDB futures

Type Aliases

Details

A vector of key value string pairs to pass to WorkloadContext::trace

SimDatabase

Rust representation of a simulated FoundationDB database

Workload

Rust representation of a FoundationDB workload