Trait foundationdb_simulation::RustWorkload

pub trait RustWorkload {
    // Required methods
    fn description(&self) -> String;
    fn setup(&'static mut self, db: SimDatabase, done: Promise);
    fn start(&'static mut self, db: SimDatabase, done: Promise);
    fn check(&'static mut self, db: SimDatabase, done: Promise);
    fn get_metrics(&self) -> Vec<Metric>;
    fn get_check_timeout(&self) -> f64;
}

RustWorkload trait provides a one to one equivalent to the C++ abstract class FDBWorkload

Required Methods

fn description(&self) -> String

Return the name or description of the workload. Primarily used for tracing.

fn setup(&'static mut self, db: SimDatabase, done: Promise)

This method is called by the tester during the setup phase. It should be used to populate the database.

Arguments

db - The simulated database.

done - A promise that should be resolved to indicate completion

fn start(&'static mut self, db: SimDatabase, done: Promise)

This method should run the actual test.

Arguments

db - The simulated database.

done - A promise that should be resolved to indicate completion

fn check(&'static mut self, db: SimDatabase, done: Promise)

This method is called when the tester completes. A workload should run any consistency/correctness tests during this phase.

Arguments

db - The simulated database.

done - A promise that should be resolved to indicate completion

fn get_metrics(&self) -> Vec<Metric>

If a workload collects metrics (like latencies or throughput numbers), these should be reported back here. The multitester (or test orchestrator) will collect all metrics from all test clients and it will aggregate them.

fn get_check_timeout(&self) -> f64

Set the check timeout for this workload.